Add `display.headers` config
New `display.headers` config to control whether SL object tables show
a header of field names at the top.

Update the default config and the config schema to reflect the new
setting.

Add long options to all commands that display these tables to allow
users to override their config on a command-by-command basis. If the
option is not present, the value of the `display.headers` config is
used.

Update commands, along with their documentation, to reflect the new
config setting.